Ignore Buildkite agent and API token from Terraform
For some reason Terraform will inform the values have changes when they
have not and will trigger a CloudFormation update. This in itself
shouldn't be an issue but when CloudFormation reports there's no changes
to be made an error is shown when it shouldn't. This can be removed when
the issue is resolved in Terraform.